requirements/base.txt
django
djangorestframework==3.11.1
djangorestframework_simplejwt
django-environ
drf-yasg
uritemplate

pylint-django